Filter and sort events by date (#37)
We don't want to show events that have already happened to the front end,
because we can't do anything with them. So now we filter the rows returned from
the database to exclude those that have already happened. As part of this work,
future events are sorted in ascending order.